Deletes JsonSchemaStrategy wrapper, promotes JsonSchemaNode/JsonSchemaParser to Conjecture.JsonSchema.Abstractions, drops OpenApi InternalsVisibleTo grant
