+++ src/Conjecture.JsonSchema.Abstractions/JsonSchemaParser.cs
@@ -2,13 +2,17 @@
 // See LICENSE.txt in the project root or https://mozilla.org/MPL/2.0/
 
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Text.Json;
 
-namespace Conjecture.JsonSchema;
+namespace Conjecture.Abstractions.JsonSchema;
 
-internal static class JsonSchemaParser
+/// <summary>Parses a JSON Schema document into a <see cref="JsonSchemaNode"/> tree.</summary>
+[EditorBrowsable(EditorBrowsableState.Never)]
+public static class JsonSchemaParser
 {
-    internal static JsonSchemaNode Parse(JsonElement root)
+    /// <summary>Parses <paramref name="root"/> into a <see cref="JsonSchemaNode"/> tree.</summary>
+    public static JsonSchemaNode Parse(JsonElement root)
     {
         Dictionary<string, JsonSchemaNode>? defs = null;
         if (root.TryGetProperty("$defs", out JsonElement defsElement))
@@ -21,85 +25,6 @@ internal static class JsonSchemaParser
         }
 
         return ParseElement(root, defs);
-    }
-
-    internal static JsonSchemaNode Resolve(JsonSchemaNode root)
-    {
-        HashSet<string> visiting = [];
-        return ResolveNode(root, root.Defs, visiting);
-    }
-
-    private static JsonSchemaNode ResolveNode(
-        JsonSchemaNode node,
-        IReadOnlyDictionary<string, JsonSchemaNode>? defs,
-        HashSet<string> visiting)
-    {
-        if (node.Ref is not null)
-        {
-            string refValue = node.Ref;
-            if (visiting.Contains(refValue))
-            {
-                return node with { Ref = null };
-            }
-
-            visiting.Add(refValue);
-            JsonSchemaNode? resolved = JsonSchemaRefResolver.ResolveRef(refValue, defs);
-            if (resolved is not null)
-            {
-                JsonSchemaNode result = ResolveNode(resolved, defs ?? resolved.Defs, visiting);
-                visiting.Remove(refValue);
-                return result;
-            }
-
-            visiting.Remove(refValue);
-        }
-
-        return node with
-        {
-            OneOf = ResolveList(node.OneOf, defs, visiting),
-            AnyOf = ResolveList(node.AnyOf, defs, visiting),
-            AllOf = ResolveList(node.AllOf, defs, visiting),
-            Properties = ResolveDict(node.Properties, defs, visiting),
-            Items = node.Items is null ? null : ResolveNode(node.Items, defs, visiting),
-        };
-    }
-
-    private static IReadOnlyList<JsonSchemaNode>? ResolveList(
-        IReadOnlyList<JsonSchemaNode>? list,
-        IReadOnlyDictionary<string, JsonSchemaNode>? defs,
-        HashSet<string> visiting)
-    {
-        if (list is null)
-        {
-            return null;
-        }
-
-        List<JsonSchemaNode> result = new(list.Count);
-        foreach (JsonSchemaNode item in list)
-        {
-            result.Add(ResolveNode(item, defs, visiting));
-        }
-
-        return result;
-    }
-
-    private static IReadOnlyDictionary<string, JsonSchemaNode>? ResolveDict(
-        IReadOnlyDictionary<string, JsonSchemaNode>? dict,
-        IReadOnlyDictionary<string, JsonSchemaNode>? defs,
-        HashSet<string> visiting)
-    {
-        if (dict is null)
-        {
-            return null;
-        }
-
-        Dictionary<string, JsonSchemaNode> result = new(dict.Count);
-        foreach (KeyValuePair<string, JsonSchemaNode> kv in dict)
-        {
-            result[kv.Key] = ResolveNode(kv.Value, defs, visiting);
-        }
-
-        return result;
     }
 
     private static JsonSchemaNode ParseElement(JsonElement element, IReadOnlyDictionary<string, JsonSchemaNode>? defs)
