+++ src/Conjecture.JsonSchema.Abstractions/JsonSchemaType.cs
@@ -1,9 +1,12 @@
 // Copyright (c) 2026 Kim Ommundsen. Licensed under the MPL-2.0.
 // See LICENSE.txt in the project root or https://mozilla.org/MPL/2.0/
 
-namespace Conjecture.JsonSchema;
+using System.ComponentModel;
+
+namespace Conjecture.Abstractions.JsonSchema;
 
 /// <summary>The JSON Schema primitive type of a schema node.</summary>
+[EditorBrowsable(EditorBrowsableState.Never)]
 public enum JsonSchemaType
 {
     /// <summary>No type constraint specified.</summary>
