+++ src/Conjecture.JsonSchema/PublicAPI.Unshipped.txt
@@ -4,8 +4,19 @@
 *REMOVED*static Conjecture.JsonSchema.JsonSchemaGenerate.extension(Conjecture.Core.Strategy!).FromJsonSchema(string! jsonSchemaText) -> Conjecture.Core.Strategy<System.Text.Json.JsonElement>!
 *REMOVED*static Conjecture.JsonSchema.JsonSchemaGenerate.extension(Conjecture.Core.Strategy!).FromJsonSchema(System.IO.FileInfo! schemaFile) -> Conjecture.Core.Strategy<System.Text.Json.JsonElement>!
 *REMOVED*static Conjecture.JsonSchema.JsonSchemaGenerate.extension(Conjecture.Core.Strategy!).FromJsonSchema(System.Text.Json.JsonElement root) -> Conjecture.Core.Strategy<System.Text.Json.JsonElement>!
+*REMOVED*Conjecture.JsonSchema.JsonSchemaType
+*REMOVED*Conjecture.JsonSchema.JsonSchemaType.None = 0 -> Conjecture.JsonSchema.JsonSchemaType
+*REMOVED*Conjecture.JsonSchema.JsonSchemaType.Null = 1 -> Conjecture.JsonSchema.JsonSchemaType
+*REMOVED*Conjecture.JsonSchema.JsonSchemaType.Boolean = 2 -> Conjecture.JsonSchema.JsonSchemaType
+*REMOVED*Conjecture.JsonSchema.JsonSchemaType.Integer = 3 -> Conjecture.JsonSchema.JsonSchemaType
+*REMOVED*Conjecture.JsonSchema.JsonSchemaType.Number = 4 -> Conjecture.JsonSchema.JsonSchemaType
+*REMOVED*Conjecture.JsonSchema.JsonSchemaType.String = 5 -> Conjecture.JsonSchema.JsonSchemaType
+*REMOVED*Conjecture.JsonSchema.JsonSchemaType.Array = 6 -> Conjecture.JsonSchema.JsonSchemaType
+*REMOVED*Conjecture.JsonSchema.JsonSchemaType.Object = 7 -> Conjecture.JsonSchema.JsonSchemaType
+*REMOVED*Conjecture.JsonSchema.JsonSchemaType.Any = 8 -> Conjecture.JsonSchema.JsonSchemaType
 Conjecture.JsonSchema.JsonSchemaStrategyExtensions
 Conjecture.JsonSchema.JsonSchemaStrategyExtensions.extension(Conjecture.Core.Strategy!)
 static Conjecture.JsonSchema.JsonSchemaStrategyExtensions.extension(Conjecture.Core.Strategy!).FromJsonSchema(string! jsonSchemaText) -> Conjecture.Core.Strategy<System.Text.Json.JsonElement>!
 static Conjecture.JsonSchema.JsonSchemaStrategyExtensions.extension(Conjecture.Core.Strategy!).FromJsonSchema(System.IO.FileInfo! schemaFile) -> Conjecture.Core.Strategy<System.Text.Json.JsonElement>!
 static Conjecture.JsonSchema.JsonSchemaStrategyExtensions.extension(Conjecture.Core.Strategy!).FromJsonSchema(System.Text.Json.JsonElement root) -> Conjecture.Core.Strategy<System.Text.Json.JsonElement>!
+static Conjecture.JsonSchema.JsonSchemaStrategyExtensions.extension(Conjecture.Core.Strategy!).FromJsonSchema(Conjecture.Abstractions.JsonSchema.JsonSchemaNode! node, int maxDepth = 5) -> Conjecture.Core.Strategy<System.Text.Json.JsonElement>!
